Split config.rs to smaller files & move config checks there
